+++ pyodide-build/pyodide_build/common.py
@@ -165,10 +165,6 @@ def _parse_package_subset(query: str | None) -> set[str]:
         packages |= CORE_PACKAGES | CORE_SCIPY_PACKAGES
         packages.discard("min-scipy-stack")
 
-    # Hack to deal with the circular dependence between soupsieve and
-    # beautifulsoup4
-    if "beautifulsoup4" in packages:
-        packages.add("soupsieve")
     packages.discard("")
     return packages
 
@@ -297,8 +293,6 @@ def get_unisolated_packages() -> list[str]:
             config = parse_package_config(pkg, check=False)
             if config.get("build", {}).get("cross-build-env", False):
                 unisolated_packages.append(config["package"]["name"])
-        # TODO: remove setuptools_rust from this when they release the next version.
-        unisolated_packages.append("setuptools_rust")
     os.environ["UNISOLATED_PACKAGES"] = json.dumps(unisolated_packages)
     return unisolated_packages