+++ src/js/load-package.ts
@@ -73,18 +73,43 @@ function _uri_to_package_name(package_uri: string): string | undefined {
   }
 }
 
+type PackageLoadMetadata = {
+  name: string;
+  channel: string;
+  depends: string[];
+  done: ResolvablePromise;
+  installPromise?: Promise<void>;
+};
+
+interface ResolvablePromise extends Promise<void> {
+  resolve: (value?: any) => void;
+  reject: (err?: Error) => void;
+}
+
+function createDonePromise(): ResolvablePromise {
+  let _resolve: (value: any) => void = () => {};
+  let _reject: (err: Error) => void = () => {};
+
+  const p: any = new Promise<void>((resolve, reject) => {
+    _resolve = resolve;
+    _reject = reject;
+  });
+
+  p.resolve = _resolve;
+  p.reject = _reject;
+  return p;
+}
+
 /**
- * Recursively add a package and its dependencies to toLoad and toLoadShared.
+ * Recursively add a package and its dependencies to toLoad.
  * A helper function for recursiveDependencies.
  * @param name The package to add
  * @param toLoad The set of names of packages to load
- * @param toLoadShared The set of names of shared libraries to load
  * @private
  */
 function addPackageToLoad(
   name: string,
-  toLoad: Map<string, string>,
-  toLoadShared: Map<string, string>
+  toLoad: Map<string, PackageLoadMetadata>
 ) {
   name = name.toLowerCase();
   if (toLoad.has(name)) {
@@ -94,11 +119,15 @@ function addPackageToLoad(
   if (!pkg_info) {
     throw new Error(`No known package with name '${name}'`);
   }
-  if (pkg_info.shared_library) {
-    toLoadShared.set(name, DEFAULT_CHANNEL);
-  } else {
-    toLoad.set(name, DEFAULT_CHANNEL);
-  }
+
+  toLoad.set(name, {
+    name: name,
+    channel: DEFAULT_CHANNEL,
+    depends: pkg_info.depends,
+    installPromise: undefined,
+    done: createDonePromise(),
+  });
+
   // If the package is already loaded, we don't add dependencies, but warn
   // the user later. This is especially important if the loaded package is
   // from a custom url, in which case adding dependencies is wrong.
@@ -107,40 +136,47 @@ function addPackageToLoad(
   }
 
   for (let dep_name of pkg_info.depends) {
-    addPackageToLoad(dep_name, toLoad, toLoadShared);
+    addPackageToLoad(dep_name, toLoad);
   }
 }
 
 /**
  * Calculate the dependencies of a set of packages
  * @param names The list of names whose dependencies we need to calculate.
- * @returns Two sets, the set of normal dependencies and the set of shared
- * dependencies
+ * @returns The map of package names to PackageLoadMetadata
  * @private
  */
 function recursiveDependencies(
   names: string[],
   errorCallback: (err: string) => void
-) {
-  const toLoad = new Map();
-  const toLoadShared = new Map();
+): Map<string, PackageLoadMetadata> {
+  const toLoad: Map<string, PackageLoadMetadata> = new Map();
   for (let name of names) {
     const pkgname = _uri_to_package_name(name);
     if (pkgname === undefined) {
-      addPackageToLoad(name, toLoad, toLoadShared);
+      addPackageToLoad(name, toLoad);
       continue;
     }
-    if (toLoad.has(pkgname) && toLoad.get(pkgname) !== name) {
+
+    const channel = name;
+
+    if (toLoad.has(pkgname) && toLoad.get(pkgname)!.channel !== channel) {
       errorCallback(
-        `Loading same package ${pkgname} from ${name} and ${toLoad.get(
-          pkgname
-        )}`
+        `Loading same package ${pkgname} from ${channel} and ${
+          toLoad.get(pkgname)!.channel
+        }`
       );
       continue;
     }
-    toLoad.set(pkgname, name);
+    toLoad.set(pkgname, {
+      name: pkgname,
+      channel: channel, // name is url in this case
+      depends: [],
+      installPromise: undefined,
+      done: createDonePromise(),
+    });
   }
-  return [toLoad, toLoadShared];
+  return toLoad;
 }
 
 //
@@ -198,7 +234,7 @@ async function downloadPackage(
 /**
  * Install the package into the file system.
  * @param name The name of the package
- * @param buffer The binary data returned by downloadPkgBuffer
+ * @param buffer The binary data returned by downloadPackage
  * @private
  */
 async function installPackage(
@@ -227,6 +263,50 @@ async function installPackage(
   });
   for (const dynlib of dynlibs) {
     await loadDynlib(dynlib, pkg.shared_library);
+  }
+}
+
+/**
+ * Download and install the package.
+ * Downloads can be done in parallel, but installs must be done for dependencies first.
+ * @param name The name of the package
+ * @param toLoad The map of package names to PackageLoadMetadata
+ * @param loaded The set of loaded package names, this will be updated by this function.
+ * @param failed The map of <failed package name, error message>, this will be updated by this function.
+ * @private
+ */
+async function downloadAndInstall(
+  name: string,
+  toLoad: Map<string, PackageLoadMetadata>,
+  loaded: Set<string>,
+  failed: Map<string, Error>
+) {
+  if (loadedPackages[name] !== undefined) {
+    return;
+  }
+
+  const pkg = toLoad.get(name)!;
+
+  try {
+    const buffer = await downloadPackage(pkg.name, pkg.channel);
+    const installPromisDependencies = pkg.depends.map((dependency) => {
+      return toLoad.has(dependency)
+        ? toLoad.get(dependency)!.done
+        : Promise.resolve();
+    });
+
+    // wait until all dependencies are installed
+    await Promise.all(installPromisDependencies);
+
+    await installPackage(pkg.name, buffer, pkg.channel);
+    loaded.add(pkg.name);
+    loadedPackages[pkg.name] = pkg.channel;
+  } catch (err: any) {
+    failed.set(name, err);
+    // We don't throw error when loading a package fails, but just report it.
+    // pkg.done.reject(err);
+  } finally {
+    pkg.done.resolve();
   }
 }
 
@@ -354,104 +434,75 @@ export async function loadPackage(
     names = [names as string];
   }
 
-  const [toLoad, toLoadShared] = recursiveDependencies(names, errorCallback);
+  const toLoad = recursiveDependencies(names, errorCallback);
 
-  for (const [pkg, uri] of [...toLoad, ...toLoadShared]) {
+  for (const [pkg, pkg_metadata] of toLoad) {
     const loaded = loadedPackages[pkg];
     if (loaded === undefined) {
       continue;
     }
     toLoad.delete(pkg);
-    toLoadShared.delete(pkg);
     // If uri is from the DEFAULT_CHANNEL, we assume it was added as a
     // dependency, which was previously overridden.
-    if (loaded === uri || uri === DEFAULT_CHANNEL) {
+    if (
+      loaded === pkg_metadata.channel ||
+      pkg_metadata.channel === DEFAULT_CHANNEL
+    ) {
       messageCallback(`${pkg} already loaded from ${loaded}`);
     } else {
       errorCallback(
-        `URI mismatch, attempting to load package ${pkg} from ${uri} ` +
+        `URI mismatch, attempting to load package ${pkg} from ${pkg_metadata.channel} ` +
           `while it is already loaded from ${loaded}. To override a dependency, ` +
           `load the custom package first.`
       );
     }
   }
 
-  if (toLoad.size === 0 && toLoadShared.size === 0) {
+  if (toLoad.size === 0) {
     messageCallback("No new packages to load");
     return;
   }
 
-  const packageNames = [...toLoad.keys(), ...toLoadShared.keys()].join(", ");
+  const packageNames = [...toLoad.keys()].join(", ");
+  const loaded = new Set<string>();
+  const failed = new Map<string, Error>();
   const releaseLock = await acquirePackageLock();
   try {
     messageCallback(`Loading ${packageNames}`);
-    const sharedLibraryLoadPromises: { [name: string]: Promise<Uint8Array> } =
-      {};
-    const packageLoadPromises: { [name: string]: Promise<Uint8Array> } = {};
-    for (const [name, channel] of toLoadShared) {
-      if (loadedPackages[name]) {
-        // Handle the race condition where the package was loaded between when
-        // we did dependency resolution and when we acquired the lock.
-        toLoadShared.delete(name);
-        continue;
-      }
-      sharedLibraryLoadPromises[name] = downloadPackage(name, channel);
-    }
-    for (const [name, channel] of toLoad) {
+    for (const [name] of toLoad) {
       if (loadedPackages[name]) {
         // Handle the race condition where the package was loaded between when
         // we did dependency resolution and when we acquired the lock.
         toLoad.delete(name);
         continue;
       }
-      packageLoadPromises[name] = downloadPackage(name, channel);
+
+      // TODO: add support for prefetching modules by awaiting on a promise right
+      // here which resolves in loadPyodide when the bootstrap is done.
+      toLoad.get(name)!.installPromise = downloadAndInstall(
+        name,
+        toLoad,
+        loaded,
+        failed
+      );
     }
 
-    const loaded: string[] = [];
-    const failed: { [name: string]: any } = {};
-    // TODO: add support for prefetching modules by awaiting on a promise right
-    // here which resolves in loadPyodide when the bootstrap is done.
-    const sharedLibraryInstallPromises: { [name: string]: Promise<void> } = {};
-    const packageInstallPromises: { [name: string]: Promise<void> } = {};
-    for (const [name, channel] of toLoadShared) {
-      sharedLibraryInstallPromises[name] = sharedLibraryLoadPromises[name]
-        .then(async (buffer) => {
-          await installPackage(name, buffer, channel);
-          loaded.push(name);
-          loadedPackages[name] = channel;
-        })
-        .catch((err) => {
-          console.warn(err);
-          failed[name] = err;
-        });
-    }
-
-    await Promise.all(Object.values(sharedLibraryInstallPromises));
-    for (const [name, channel] of toLoad) {
-      packageInstallPromises[name] = packageLoadPromises[name]
-        .then(async (buffer) => {
-          await installPackage(name, buffer, channel);
-          loaded.push(name);
-          loadedPackages[name] = channel;
-        })
-        .catch((err) => {
-          console.warn(err);
-          failed[name] = err;
-        });
-    }
-    await Promise.all(Object.values(packageInstallPromises));
+    await Promise.all(
+      Array.from(toLoad.values()).map(({ installPromise }) => installPromise)
+    );
 
     Module.reportUndefinedSymbols();
-    if (loaded.length > 0) {
-      const successNames = loaded.join(", ");
+    if (loaded.size > 0) {
+      const successNames = Array.from(loaded).join(", ");
       messageCallback(`Loaded ${successNames}`);
     }
-    if (Object.keys(failed).length > 0) {
-      const failedNames = Object.keys(failed).join(", ");
+
+    if (failed.size > 0) {
+      const failedNames = Array.from(failed.keys()).join(", ");
       messageCallback(`Failed to load ${failedNames}`);
-      for (const [name, err] of Object.entries(failed)) {
-        console.warn(`The following error occurred while loading ${name}:`);
-        console.error(err);
+      for (const [name, err] of failed) {
+        errorCallback(`The following error occurred while loading ${name}:`);
+        errorCallback(err.message);
       }
     }